// src/data/api/generated/models/shareAccess.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Lupira Tasks API
 * Task and command processing backend for Lupira. Authenticate with a Bearer token issued by the OIDC provider (Authentik).
 * OpenAPI spec version: v1
 */

export type ShareAccess = typeof ShareAccess[keyof typeof ShareAccess];


export const ShareAccess = {
  Read: 'Read',
  ReadWrite: 'ReadWrite',
} as const;